+++ .idea/artifacts/tmdb_jvm.xml
@@ -1,0 +1,6 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="tmdb-jvm">
+    <output-path>$PROJECT_DIR$/api/tmdb/build/libs</output-path>
+    <root id="archive" name="tmdb-jvm.jar" />
+  </artifact>
+</component>
